Move more stuff to the extension
